api/requirements.txt
fastapi==0.111.1
google_api_python_client==2.137.0
httpx==0.27.0
protobuf==5.27.2
uvicorn==0.30.1
pytz==2022.7.1